+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/sider/runners/compare/0.28.1...HEAD)
 
+- Retry git-fetch(1) and git-checkout(1) [#1214](https://github.com/sider/runners/pull/1214)
+
 ## 0.28.1
 
 [Full diff](https://github.com/sider/runners/compare/0.28.0...0.28.1)